kompletni uprava cele stranky

+++ server/controllers/clothing.js
@@ -1,12 +1,14 @@
+require('dotenv').config();
 const Clothing = require("../models/clothing");
+const ADMIN_PASSWORD = process.env.ADMIN_PASSWORD;
 
 exports.getAllClothing = async (req, res, next) => {
   try {
     let queryName = req.query.name;
-
+    console.log(queryName);
     let data;
     if (queryName) 
-      data = await Clothing.find({name: queryName});
+      data = await Clothing.find({ name: { $regex: queryName, $options: 'i' } });
     else 
       data = await Clothing.find();
 
@@ -16,13 +18,13 @@ exports.getAllClothing = async (req, res, next) => {
         payload: data,
       });
     }
-    res.status(404).send({
-      message: "Clothing not found",
-    });
+    res.status(404).send({ message: "Clothing not found" });
   } catch (err) {
     res.status(500).send(err);
   }
 };
+
+
 exports.getClothingById = async (req, res, next) => {
   try {
     const data = await Clothing.findById(req.params.id);
@@ -32,15 +34,19 @@ exports.getClothingById = async (req, res, next) => {
         payload: data,
       });
     }
-    res.status(404).send({
-      message: "Clothing not found",
-    });
+    res.status(404).send({ message: "Clothing not found" });
   } catch (err) {
     res.status(500).send(err);
   }
 };
+
 exports.createClothing = async (req, res, next) => {
   try {
+    const passwordReq = req.body.password;
+    if (ADMIN_PASSWORD !== passwordReq) {
+      return res.status(403).send({ message: "Incorrect password" });
+    }
+
     const data = new Clothing({
       name: req.body.name,
       material: req.body.material,
@@ -50,6 +56,7 @@ exports.createClothing = async (req, res, next) => {
       price: req.body.price,
       image: req.body.image,
     });
+
     const result = await data.save();
     if (result) {
       return res.status(201).send({
@@ -57,15 +64,19 @@ exports.createClothing = async (req, res, next) => {
         payload: result,
       });
     }
-    res.status(500).send({
-      message: "not found",
-    });
+    res.status(500).send({ message: "Not created" });
   } catch (err) {
     res.status(500).send(err);
   }
 };
+
 exports.updateClothing = async (req, res, next) => {
   try {
+    const passwordReq = req.body.password;
+    if (ADMIN_PASSWORD !== passwordReq) {
+      return res.status(403).send({ message: "Incorrect password" });
+    }
+
     const data = {
       name: req.body.name,
       material: req.body.material,
@@ -75,6 +86,7 @@ exports.updateClothing = async (req, res, next) => {
       price: req.body.price,
       image: req.body.image,
     };
+
     const result = await Clothing.findByIdAndUpdate(req.params.id, data);
     if (result) {
       return res.status(200).send({
@@ -82,15 +94,19 @@ exports.updateClothing = async (req, res, next) => {
         payload: result,
       });
     }
-    res.status(500).send({
-      message: "Clothing not updated",
-    });
+    res.status(500).send({ message: "Clothing not updated" });
   } catch (err) {
     res.status(500).send(err);
   }
 };
+
 exports.deleteClothing = async (req, res, next) => {
   try {
+    const passwordReq = req.body.password;
+    if (ADMIN_PASSWORD !== passwordReq) {
+      return res.status(403).send({ message: "Incorrect password" });
+    }
+
     const result = await Clothing.findByIdAndDelete(req.params.id);
     if (result) {
       return res.status(200).send({
@@ -98,9 +114,7 @@ exports.deleteClothing = async (req, res, next) => {
         payload: result,
       });
     }
-    res.status(500).send({
-      message: "Clothing not deleted",
-    });
+    res.status(500).send({ message: "Clothing not deleted" });
   } catch (err) {
     res.status(500).send(err);
   }
